pyproject.toml
[project]
name = "strouhal-galloping-surrogates"
version = "0.1.0"
description = "ML surrogates for VIV/galloping stability analysis of triangular bluff bodies"
authors = [
    {name = "Ming Zhang"}
]
requires-python = ">=3.11,<3.13"
dependencies = [
    # Core scientific computing
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.10.0",
    "h5py>=3.8.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "pyarrow>=12.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.23.0",
    "notebook>=7.0.0",
    "lightgbm>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
